src/requirements.txt
# torch, torchaudio, silero-vad removed; using webrtcvad-wheels for VAD
webrtcvad-wheels==2.0.14
numpy<2.0
soundfile==0.12.1
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==1.10.13
aiohttp==3.9.1
redis==5.0.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
httpx==0.27.2  # Pin to 0.27.2 to avoid 'proxies' argument error
livekit==1.0.0
pyjwt==2.8.0
# aioredis removed (deprecated, use redis>=4.2.0)